About page and optimizations done

+++ db/migrate/20200520100447_create_specific_menu_items.rb
@@ -1,8 +1,8 @@
 class CreateSpecificMenuItems < ActiveRecord::Migration[6.0]
   def change
     create_table :specific_menu_items do |t|
-      t.integer :menu_id
-      t.integer :menu_item_id
+      t.bigint :menu_id
+      t.bigint :menu_item_id
     end
   end
 end
